appended model tests
